requirements.txt
sqlite3~=0.0.0
PyYAML~=5.4.1
tabulate~=0.8.9
setuptools~=56.2.0
pytest~=6.2.4
graphviz~=0.16
click~=8.0.1